Close #14, close #13.
A few things here. First, `eesim` wasn't passing all the custom arguments through to later functions, which is why there was an error for the custom exposure example. For the `baseline` bug, the problem was that `baseline` came through as a numerical vector in this case, which the `create_lambda` is expecting a dataframe with a column also named `baseline`. I added a line to the function to check if `baseline` is a dataframe and, if so, to pull out just the `baseline` column for the `baseline` object. This isn't terribly elegant, but it should get the job done.